+++ esphome/core/hal/hal_esp32.h
@@ -42,6 +42,8 @@ __attribute__((always_inline)) inline void delayMicroseconds(uint32_t us) { dela
 __attribute__((always_inline)) inline void arch_feed_wdt() { esp_task_wdt_reset(); }
 __attribute__((always_inline)) inline uint32_t arch_get_cpu_cycle_count() { return esp_cpu_get_cycle_count(); }
 
+void arch_init();
+
 }  // namespace esphome
 
 #endif  // USE_ESP32
