+++ esphome/core/hal/hal_esp8266.h
@@ -58,6 +58,9 @@ __attribute__((always_inline)) inline uint16_t progmem_read_uint16(const uint16_
 // NOLINTNEXTLINE(readability-identifier-naming)
 __attribute__((always_inline)) inline void delayMicroseconds(uint32_t us) { delay_microseconds_safe(us); }
 __attribute__((always_inline)) inline void arch_feed_wdt() { system_soft_wdt_feed(); }
+__attribute__((always_inline)) inline void arch_init() {}
+
+uint32_t arch_get_cpu_cycle_count();
 
 }  // namespace esphome
