+++ esphome/core/hal/hal_libretiny.h
@@ -91,6 +91,7 @@ __attribute__((always_inline)) inline uint64_t millis_64() { return Millis64Impl
 void delayMicroseconds(uint32_t us);  // NOLINT(readability-identifier-naming)
 void arch_feed_wdt();
 uint32_t arch_get_cpu_cycle_count();
+void arch_init();
 
 }  // namespace esphome
