+++ esphome/core/hal/hal_rp2040.h
@@ -38,6 +38,7 @@ __attribute__((always_inline)) inline uint64_t millis_64() { return micros_to_mi
 void delayMicroseconds(uint32_t us);  // NOLINT(readability-identifier-naming)
 void arch_feed_wdt();
 uint32_t arch_get_cpu_cycle_count();
+void arch_init();
 
 }  // namespace esphome
