+++ esphome/core/hal/hal_zephyr.h
@@ -22,6 +22,7 @@ uint64_t millis_64();
 void delayMicroseconds(uint32_t us);  // NOLINT(readability-identifier-naming)
 void arch_feed_wdt();
 uint32_t arch_get_cpu_cycle_count();
+void arch_init();
 
 }  // namespace esphome
